option-requests

+++ src/app/success/page.tsx
@@ -4,6 +4,9 @@ import { useEffect, useState } from "react";
 import Link from "next/link";
 import Image from "next/image";
 import LogoutIcon from "@mui/icons-material/Logout";
+import { addBooking } from "@/api";
+import { useMutation } from "@tanstack/react-query";
+import { FormValues } from "@/types/interfaces";
 
 const BgnImg = () => {
   return (
@@ -28,12 +31,31 @@ const BgnImg = () => {
 
 const Success = () => {
   const [timer, setTimer] = useState(5);
+  const mutation = useMutation({
+    mutationFn: (booking: FormValues) => addBooking(booking),
+    onSuccess: (data) => {
+      console.log(data);
+    },
+    onError: (error: any) => {
+      console.log(error);
+    },
+  });
 
   useEffect(() => {
+    const objectBooking = localStorage.getItem("form");
+
+    if (!objectBooking) {
+      return;
+    }
+    const booking = JSON.parse(objectBooking);
+    console.log(booking);
+    mutation.mutate(booking);
+
     localStorage.clear();
     const interval = setInterval(() => {
       setTimer((prevTimer) => {
         if (prevTimer <= 1) {
+          clearInterval(interval);
           window.location.href = "/";
           return 0;
         }
@@ -42,7 +64,13 @@ const Success = () => {
     }, 1000);
 
     return () => clearInterval(interval);
-  }, []);
+  }, [mutation]);
+
+  useEffect(() => {
+    if (timer <= 0) {
+      window.location.href = "/";
+    }
+  }, [timer]);
 
   return (
     <div className="relative h-[calc(100vh-84px)] md:h-[calc(100vh-96px)] lg:h-screen  flex flex-col justify-between ">
